Refactor processTransaction(), wallet_connect_approve_action() and wallet_connect_approve_action() to better manage chainId
In wallet_connect_approve_action() and wallet_connect_approve_action(): select Network instead of chainId

In processTransaction():
Allow user to change chainId (via wallet_connect_approve_action) before signing a transaction.
This is because some Dapps do not seem to switch chain correctly.
For example, MyEtherWallet does not provide chainId when submitting a tx to WalletConnect, so the Satochip-Bridge has no way of knowing or setting the correct chainId.
In this case, it is the user responsability to check and eventually modify the chainId so that the Dapp and Bridge chainId are compatible.

Some Apps (e.g. Metamask) provide the chainId when submitting a transaction through WalletConnect (even for legacy tx)